+++ tests/test_sce.py
@@ -24,7 +24,7 @@ def rosenbrock_args(x, a, b=100):
     Rosenbrock, global optimum: f(a, a^2) = 0.0.
 
     """
-    f = (a - x[0])**2 + b*(x[1] - x[0]**2)**2
+    f = (a - x[0])**2 + b * (x[1] - x[0]**2)**2
 
     return f
 
@@ -34,7 +34,7 @@ def rosenbrock_kwargs(x, a=1, b=100):
     Rosenbrock, global optimum: f(a, a^2) = 0.0.
 
     """
-    f = (a - x[0])**2 + b*(x[1] - x[0]**2)**2
+    f = (a - x[0])**2 + b * (x[1] - x[0]**2)**2
 
     return f
 
@@ -164,6 +164,12 @@ class TestSCESolver(unittest.TestCase):
         assert_almost_equal(result.x, self.rosenx, decimal=3)
         # degenerated bounds
         x0 = [0.999, 0.5, 0.1]
+        lb = [2., 0., 0.]
+        ub = [2., 2., 2.]
+        rosenx = [x0[0], 1., 1.]
+        result = sce(rosen, x0, lb, ub)
+        assert_almost_equal(result.x, rosenx, decimal=2)
+        x0 = [0.999, 0.5, 0.1]
         lb = [3., 0., 0.]
         ub = [2., 2., 2.]
         rosenx = [x0[0], 1., 1.]
@@ -204,13 +210,14 @@ class TestSCESolver(unittest.TestCase):
         assert_equal(result.fun, rosen(result.x))
         result = sce(rosen, self.x0, self.lb, self.ub, printit=2, maxn=10)
         assert_equal(result.fun, rosen(result.x))
-        result = sce(rosen, [0.5]*5, [0.]*5, [2.]*5, printit=1, pcento=10)
+        result = sce(rosen, [0.5] * 5, [0.] * 5, [2.] * 5, printit=1,
+                     pcento=10)
         print(result.message)
         assert_equal(result.fun, rosen(result.x))
 
         # sampling
         result = sce(rosen, self.x0, self.lb, self.ub,
-                     sampling=['half-open']*len(self.lb))
+                     sampling=['half-open'] * len(self.lb))
         assert_almost_equal(result.x, self.rosenx, decimal=3)
         smpls = ['half-open', 'right-half-open', 'left-half-open', 'open',
                  'log']