+++ src/pyjams/__init__.py
@@ -85,6 +85,7 @@ History
    * v1.29, added functions for converting humidity in air,
      Jan 2023, Matthias Cuntz
    * v1.30, hvplot in mcplot, May 2023, Matthias Cuntz
+   * v1.31, added pack and unpack, Jun 2023, Matthias Cuntz
 
 """
 # version, author